refactor: split studio shell and inspector modules
